chore: 홈 및 프로필 관련 모델/리포지토리 파일 구조 정리
