+++ app/src/main/java/com/solux/moro/data/model/CommentItem.kt
@@ -1,0 +1,8 @@
+package com.solux.moro.data.model
+
+data class CommentItem(
+    val id: String,
+    val userNickname: String,
+    val content: String,
+    val createdAt: Long
+)
