+++ app/src/main/java/com/solux/moro/data/repository/UserRepository.kt
@@ -1,7 +1,10 @@
-package com.solux.moro.ui.profile
+package com.solux.moro.data.repository
 
 import com.solux.moro.core.designsystem.theme.MoroThemeType
-import com.solux.moro.ui.home.FeedItem
+import com.solux.moro.data.model.FeedItem
+import com.solux.moro.data.model.UserColorPalette
+import com.solux.moro.data.model.User
+import com.solux.moro.data.model.UserStats
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.StateFlow
