+++ app/src/main/java/com/solux/moro/ui/home/FeedViewModel.kt
@@ -2,6 +2,8 @@ package com.solux.moro.ui.home
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.solux.moro.data.model.FeedItem
+import com.solux.moro.data.repository.FeedRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import jakarta.inject.Inject
 import kotlinx.coroutines.flow.SharingStarted
